poa_agent v0.1.0 POAAgent.Plugins.Transfers.WebSocket.Primus

Functions

data_received(label, metric_type, data, state)

In this callback is called when a Collector sends data to this Transfer.

Regarding the parameters, label identifies the Collector, metric_type is the metric type in String format, it is used in order to know which kind of data we are sending, the data is the real data received, and the state is the Collector’s state

It must return {:ok, state}.

Callback implementation for POAAgent.Plugins.Transfer.data_received/4.

handle_message(arg1, state)

In this callback is called when the Transfer process receives an erlang message.

It must return {:ok, state}.

Callback implementation for POAAgent.Plugins.Transfer.handle_message/2.

init_transfer(configuration)

A callback executed when the Transfer Plugin starts. The argument is retrieved from the configuration file when the Transfer is defined It must return {:ok, state}, that state will be keept as in GenServer and can be retrieved in the data_received/2 function.

Callback implementation for POAAgent.Plugins.Transfer.init_transfer/1.

terminate(_)

This callback is called just before the Process goes down. This is a good place for closing connections.

Callback implementation for POAAgent.Plugins.Transfer.terminate/1.